商品削除機能の修正

+++ app/controllers/items_controller.rb
@@ -1,5 +1,5 @@
 class ItemsController < ApplicationController
-  before_action :authenticate_user!, only: :new
+  before_action :authenticate_user!, only: [:new, :destroy]
   before_action :set_item, only: [:edit, :show, :update]
   before_action :move_to_signIn, except: [:index, :show]
   before_action :move_to_index, only: [:edit]
@@ -26,6 +26,7 @@ class ItemsController < ApplicationController
 
   def edit
     return unless @item.order.present?
+
     redirect_to action: :index
   end
 
@@ -39,6 +40,8 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
+    return unless item.user_id == current_user.id
+
     item.destroy
     redirect_to root_path
   end
@@ -51,11 +54,13 @@ class ItemsController < ApplicationController
 
   def move_to_signIn
     return if user_signed_in?
+
     redirect_to new_user_session_path
   end
 
   def move_to_index
     return if current_user.id == @item.user_id
+
     redirect_to action: :index
   end
